refactor(ui): extract ItemRenderer and share socket and gem building
ItemRenderer moves out of gear_picker.tsx into its own file. The JSX and
every class name are unchanged, so no SCSS is affected. What changes:

- clear(slot) + update(item) become one render(item | null). The two phases
  stay as private reset()/apply() because apply() deliberately does not
  clear stale state on its negative paths -- reset() is what does that. The
  slot moves to the constructor, absorbing the empty-state name and quality
  reset that ItemPicker used to do by hand.
- The handles nobody outside reads are now private: ilvlElem,
  nameContainerElem, socketsContainerElem, abortController and signal.
- createItemSockets in utils.tsx replaces two near-verbatim copies of the
  socket-building loop. It returns the Disposable from professionChangeEmitter
  instead of discarding it, so the caller can dispose it before the next
  render; previously both copies leaked a listener on every gear change.
- setGemInContainer replaces the gem fill that selector_modal had
  re-implemented after reaching into createGemContainer's output with
  querySelector.
- The enchant and tinker labels share one code path, which fixes a
  copy-paste slip where the tinker's item branch wrote enchantElem.href.
- The async enchant description and tooltip lookups now honour the abort
  signal, so a slow response for a previous item cannot land on the current
  one.

Verified against master: all 38 gear cells render identically (name, ilvl,
icon, wowhead dataset, sockets, gem icons, enchant, tinker, reforge,
quality), and the empty-slot state matches field for field.
