+++ ui/core/components/gear_picker/selector_modal.tsx
@@ -16,7 +16,7 @@ import { mod, randomUUID, sanitizeId } from '../../utils';
 import { BaseModal } from '../base_modal';
 import GearPicker from './gear_picker';
 import ItemList, { GearData, ItemData, ItemListType } from './item_list';
-import { createGemContainer, getEmptySlotIconUrl } from './utils';
+import { createGemContainer, getEmptySlotIconUrl, setGemInContainer } from './utils';
 import i18n from '../../../i18n/config';
 
 export enum SelectorModalTabs {
@@ -414,24 +414,10 @@ export default class SelectorModal extends BaseModal {
 					tabButton.appendChild(gemContainer);
 					tabButton.classList.add('selector-modal-tab-gem');
 
-					const gemElem = tabButton.querySelector<HTMLElement>('.gem-icon')!;
 					const emptySocketUrl = getEmptyGemSocketIconUrl(socketColor);
 
 					const updateGemIcon = () => {
-						const equippedItem = gearData.getEquippedItem();
-						const gem = equippedItem?.gems[socketIdx];
-
-						if (gem) {
-							gemElem.classList.remove('hide');
-							ActionId.fromItemId(gem.id)
-								.fill()
-								.then(filledId => {
-									gemElem.setAttribute('src', filledId.iconUrl);
-								});
-						} else {
-							gemElem.classList.add('hide');
-							gemElem.setAttribute('src', emptySocketUrl);
-						}
+						setGemInContainer(gemContainer, gearData.getEquippedItem()?.gems[socketIdx] ?? null, emptySocketUrl);
 					};
 
 					gearData.changeEvent.on(updateGemIcon);
