+++ ui/core/components/gear_picker/utils.tsx
@@ -1,9 +1,10 @@
-import { ref } from 'tsx-vanilla';
-
+import { Player } from '../../player';
 import { GemColor, ItemSlot } from '../../proto/common';
 import { UIGem as Gem } from '../../proto/ui';
 import { ActionId } from '../../proto_utils/action_id';
+import { EquippedItem } from '../../proto_utils/equipped_item';
 import { getEmptyGemSocketIconUrl } from '../../proto_utils/gems';
+import { Disposable } from '../../typed_event';
 
 const emptySlotIcons: Record<ItemSlot, string> = {
 	[ItemSlot.ItemSlotHead]: '/mop/assets/item_slots/head.jpg',
@@ -31,23 +32,51 @@ export const createNameDescriptionLabel = (nameDesc: string) => {
 	return <small className="heroic-label">({nameDesc})</small>;
 };
 
+// Points the gem icon inside a container built by createGemContainer at `gem`, or hides it when
+// the socket is empty. Resolves with the filled ActionId so callers can wire up extra state.
+export const setGemInContainer = async (container: HTMLElement, gem: Gem | null, emptySocketIconUrl: string): Promise<ActionId | null> => {
+	const gemIconElem = container.querySelector<HTMLImageElement>('.gem-icon')!;
+	if (!gem) {
+		gemIconElem.classList.add('hide');
+		gemIconElem.src = emptySocketIconUrl;
+		return null;
+	}
+
+	gemIconElem.classList.remove('hide');
+	const filledId = await ActionId.fromItemId(gem.id).fill();
+	gemIconElem.src = filledId.iconUrl;
+	return filledId;
+};
+
 export const createGemContainer = (socketColor: GemColor, gem: Gem | null, index: number) => {
-	const gemIconElem = ref<HTMLImageElement>();
-	const gemContainerElem = ref<HTMLAnchorElement>();
 	const gemContainer = (
-		<a ref={gemContainerElem} className="gem-socket-container" href="javascript:void(0)" dataset={{ socketIdx: index }}>
-			<img ref={gemIconElem} className={`gem-icon ${!gem ? 'hide' : ''}`} />
+		<a className="gem-socket-container" href="javascript:void(0)" dataset={{ socketIdx: index }}>
+			<img className="gem-icon hide" />
 			<img className="socket-icon" src={getEmptyGemSocketIconUrl(socketColor)} />
 		</a>
-	);
+	) as HTMLAnchorElement;
 
-	if (gem) {
-		ActionId.fromItemId(gem.id)
-			.fill()
-			.then(filledId => {
-				filledId.setWowheadHref(gemContainerElem.value!);
-				gemIconElem.value!.src = filledId.iconUrl;
-			});
-	}
-	return gemContainer as HTMLAnchorElement;
+	setGemInContainer(gemContainer, gem, getEmptyGemSocketIconUrl(socketColor)).then(filledId => filledId?.setWowheadHref(gemContainer));
+	return gemContainer;
+};
+
+// Builds the gem sockets for an item, including the extra socket that Blacksmithing grants.
+// The profession subscription is handed back rather than swallowed: the caller owns it and must
+// dispose it before the next render, otherwise listeners pile up on every gear change.
+export const createItemSockets = (player: Player<any>, item: EquippedItem): { elems: HTMLAnchorElement[]; professionSubscription: Disposable | null } => {
+	let professionSubscription: Disposable | null = null;
+
+	const elems = item.allSocketColors().map((socketColor, gemIdx) => {
+		const gemContainer = createGemContainer(socketColor, item.gems[gemIdx], gemIdx);
+		if (gemIdx === item.numPossibleSockets - 1 && item.couldHaveExtraSocket()) {
+			const updateProfession = () => {
+				gemContainer.classList[player.isBlacksmithing() ? 'remove' : 'add']('hide');
+			};
+			professionSubscription = player.professionChangeEmitter.on(updateProfession);
+			updateProfession();
+		}
+		return gemContainer;
+	});
+
+	return { elems, professionSubscription };
 };
