+++ ui/core/components/individual_sim_ui/bulk/bulk_item_picker.tsx
@@ -8,7 +8,7 @@ import { EquippedItem } from '../../../proto_utils/equipped_item';
 import { getEligibleItemSlots } from '../../../proto_utils/utils';
 import { TypedEvent } from '../../../typed_event';
 import { Component } from '../../component';
-import { ItemRenderer } from '../../gear_picker/gear_picker';
+import { ItemRenderer } from '../../gear_picker/item_renderer';
 import { GearData } from '../../gear_picker/item_list';
 import { SelectorModalTabs } from '../../gear_picker/selector_modal';
 import { BulkTab } from '../bulk_tab';
@@ -37,7 +37,7 @@ export default class BulkItemPicker extends Component {
 		this.bulkSlot = bulkSlot;
 		this.index = index;
 		this.item = item;
-		this.itemElem = new ItemRenderer(parent, this.rootElem, simUI.player);
+		this.itemElem = new ItemRenderer(parent, this.rootElem, simUI.player, getEligibleItemSlots(item.item)[0]);
 		this.abortController = new AbortController();
 		this.signal = this.abortController.signal;
 
@@ -67,8 +67,7 @@ export default class BulkItemPicker extends Component {
 	}
 
 	setItem(newItem: EquippedItem) {
-		this.itemElem.clear(ItemSlot.ItemSlotHead);
-		this.itemElem.update(newItem);
+		this.itemElem.render(newItem);
 		this.item = newItem;
 		this.setupHandlers();
 	}
